backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv==1.0.0
